RavenDB-19168: Fixed edge cases on unused types like sbyte and [u]short.

+++ src/Voron/Data/CompactTrees/CompactTree.cs
@@ -784,7 +784,7 @@ namespace Voron.Data.CompactTrees
             if (state.LastSearchPosition >= 0) // update
             {
                 GetValuePointer(ref state, state.LastSearchPosition, out var b);
-                ZigZagEncoding.Decode<int>(b, out var len);
+                ZigZagEncoding.Decode<long>(b, out var len);
 
                 if (len == valueBufferLength)
                 {
